Prove that match is refined by derivative-match
For this, it turned out to be better to rewrite match and split as
(perhaps mutually) recursive functions, so we can replace split by its
specification in proofs about match. Also, we had to change the
semantics of Fail so it is only allowed when there is a proof that the
result is not inhabited.