Change the definition of the invariant for recursion
This would be one potential way to explain how to give semantics for
recursion: when we perform a recursive call, we must prove the
precondition holds on the input, and for all potential output (i.e. such
that the postcondition holds), the remainder must be correct.  If there
is no recursive call, then we need to show the postcondition is
satisfied by the output. For each thing we need to prove, we are allowed
to assume the precondition on the original call is true.

The advantage of this change is that I didn't change too much, but I
prefer another definition, which I will add in the next commit.